[Resolve #37] Make separate controller for files and file statuses endpoints
