+++ src/main/java/oleksandrdiachenko/pricechecker/controller/PriceRestController.java
@@ -3,51 +3,33 @@ package oleksandrdiachenko.pricechecker.controller;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import oleksandrdiachenko.pricechecker.model.PriceCheckParameter;
-import oleksandrdiachenko.pricechecker.model.entity.File;
-import oleksandrdiachenko.pricechecker.service.FileService;
-import oleksandrdiachenko.pricechecker.service.FileStatusService;
 import oleksandrdiachenko.pricechecker.service.QueueService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.validation.constraints.Min;
-import java.util.Collections;
-import java.util.Optional;
-
-import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 @RestController
 @Validated
 @Slf4j
+@RequestMapping("/api/pricecheck")
 class PriceRestController {
 
     private final QueueService queueService;
-    private final FileStatusService fileStatusService;
-    private final FileService fileService;
 
     @Autowired
-    public PriceRestController(QueueService queueService, FileStatusService fileStatusService, FileService fileService) {
+    public PriceRestController(QueueService queueService) {
         this.queueService = queueService;
-        this.fileStatusService = fileStatusService;
-        this.fileService = fileService;
-    }
-
-    @GetMapping("/api/pricecheck/file/{id}")
-    public @ResponseBody
-    ResponseEntity<?> getFileById(@PathVariable(value = "id") long id) {
-        Optional<File> fileOptional = fileService.findById(id);
-        if (fileOptional.isPresent()) {
-            return ResponseEntity.ok(fileOptional.get().getFile());
-        }
-        return new ResponseEntity<>(
-                new ErrorResponse<>(Collections.singletonList("File with id: " + id + " not found")), NOT_FOUND);
     }
 
     @SneakyThrows
-    @PostMapping(value = "/api/pricecheck")
+    @PostMapping()
     public ResponseEntity<?> acceptFile(@RequestParam("file") MultipartFile file,
                                         @RequestParam("urlIndex") @Min(value = 1, message = "Url column should be greater than 0") int urlIndex,
                                         @RequestParam("insertIndex") @Min(value = 1, message = "Insert column should be greater than 0") int insertIndex) {
@@ -56,10 +38,4 @@ class PriceRestController {
         queueService.addToQueue(parameter);
         return ResponseEntity.accepted().build();
     }
-
-    @GetMapping(value = "/api/pricecheck/filestatuses")
-    public ResponseEntity<?> getFileStatuses() {
-        return ResponseEntity.ok(fileStatusService.findAll());
-    }
 }
-
