+++ src/test/java/oleksandrdiachenko/pricechecker/controller/PriceRestControllerTest.java
@@ -1,15 +1,9 @@
 package oleksandrdiachenko.pricechecker.controller;
 
 import oleksandrdiachenko.pricechecker.model.PriceCheckParameter;
-import oleksandrdiachenko.pricechecker.model.entity.File;
-import oleksandrdiachenko.pricechecker.model.entity.FileStatus;
-import oleksandrdiachenko.pricechecker.model.entity.Status;
-import oleksandrdiachenko.pricechecker.service.FileService;
-import oleksandrdiachenko.pricechecker.service.FileStatusService;
 import oleksandrdiachenko.pricechecker.service.QueueService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -17,13 +11,9 @@ import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 
-import java.time.LocalDateTime;
-import java.util.Collections;
-import java.util.Optional;
-
 import static org.apache.commons.lang3.StringUtils.EMPTY;
-import static org.mockito.Mockito.*;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.verify;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.multipart;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -35,29 +25,21 @@ public class PriceRestControllerTest {
     public static final String XLS_CONTENT_TYPE = "application/vnd.ms-excel";
     public static final String XLSX_CONTENT_TYPE = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet";
     public static final String PRICECHECK = "/api/pricecheck";
-    public static final String FILESTATUSES = "/api/pricecheck/filestatuses";
-    public static final String FILE_BY_ID = "/api/pricecheck/file/";
     public static final String FILE = "file";
     public static final String URL_INDEX_PARAM = "urlIndex";
     public static final String INSERT_INDEX_PARAM = "insertIndex";
-    private static String FILENAME = "filename";
-    private static byte[] BYTES = new byte[]{1, 2, 3};
-    private static int URL_COLUMN = 1;
-    private static int URL_INDEX = 0;
-    private static int INSERT_COLUMN = 2;
-    private static int INSERT_INDEX = 1;
+    private static final String FILENAME = "filename";
+    private static final byte[] BYTES = new byte[]{1, 2, 3};
+    private static final int URL_COLUMN = 1;
+    private static final int URL_INDEX = 0;
+    private static final int INSERT_COLUMN = 2;
+    private static final int INSERT_INDEX = 1;
 
     @Autowired
     private MockMvc mvc;
 
     @MockBean
     private QueueService queueService;
-    @MockBean
-    private FileStatusService fileStatusService;
-    @MockBean
-    private FileService fileService;
-    @Mock
-    private File file;
 
     @Test
     void shouldReturnOkAndCallServiceWhenFileIsXls() throws Exception {
@@ -140,47 +122,5 @@ public class PriceRestControllerTest {
                 .andExpect(content().json("{\"errors\":[\"Url column should be greater than 0\"," +
                         "\"Insert column should be greater than 0\"]}"));
         verify(queueService, never()).addToQueue(new PriceCheckParameter(FILENAME, URL_INDEX, INSERT_INDEX, BYTES));
-    }
-
-    @Test
-    void shouldReturnListWithOneFileStatusesWhenServiceReturnListWithOneFileStatus() throws Exception {
-        when(fileStatusService.findAll()).thenReturn(Collections.singletonList(
-                new FileStatus(1, FILENAME, Status.COMPLETED.name(), 2,
-                        LocalDateTime.of(2020, 1, 6, 5, 18, 20))));
-
-        mvc.perform(get(FILESTATUSES))
-                .andExpect(status().isOk())
-                .andExpect(content().json("[{\"id\":1,\"name\":\"filename\"," +
-                        "\"status\":\"COMPLETED\",\"fileId\":2,\"acceptedTime\":\"2020-01-06T05:18:20\"}]"));
-    }
-
-    @Test
-    void shouldReturnEmptyListWhenServiceReturnEmptyList() throws Exception {
-        when(fileStatusService.findAll()).thenReturn(Collections.emptyList());
-
-        mvc.perform(get(FILESTATUSES))
-                .andExpect(status().isOk())
-                .andExpect(content().json("[]"));
-    }
-
-    @Test
-    void shouldReturnBytesWhenFileExist() throws Exception {
-        long fileId = 1;
-        when(fileService.findById(fileId)).thenReturn(Optional.of(file));
-        when(file.getFile()).thenReturn(BYTES);
-
-        mvc.perform(get(FILE_BY_ID + fileId))
-                .andExpect(status().isOk())
-                .andExpect(content().bytes(file.getFile()));
-    }
-
-    @Test
-    void shouldReturnNotFoundWhenFileIsNotExist() throws Exception {
-        long fileId = -1;
-        when(fileService.findById(fileId)).thenReturn(Optional.empty());
-
-        mvc.perform(get(FILE_BY_ID + fileId))
-                .andExpect(status().isNotFound())
-                .andExpect(content().json("{\"errors\":[\"File with id: -1 not found\"]}"));
     }
 }
